Create 'useMetaEntityProp' hook
